requirements.txt
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astropy==4.3.1
attrs==22.1.0
audioread==2.1.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
decorator==5.1.1
Flask==2.1.2
Flask-Cors==3.0.10
flask-restx==0.5.1
fonttools==4.34.4
greenlet==1.1.2
greenletio==0.9.0
h11==0.13.0
hydra-core==1.2.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.9.1
kiwisolver==1.4.3
librosa==0.9.2
llvmlite==0.38.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mysql-connector-python==8.0.29
numba==0.55.2
numpy==1.21.6
omegaconf==2.2.2
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
pooch==1.6.0
protobuf==4.21.2
pycparser==2.21
pyerfa==2.0.0.1
PyJWT==2.4.0
PyMySQL==1.0.2
pymysql-pool==0.3.7
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-Levenshtein==0.12.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
resampy==0.3.1
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
six==1.16.0
SoundFile==0.10.3.post1
SQLAlchemy==1.4.39
threadpoolctl==3.1.0
torch==1.12.0
torchaudio==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.10
uvicorn==0.18.3
Werkzeug==2.1.2
zipp==3.8.0
